% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kim.t.R
\name{kim.t}
\alias{kim.t}
\title{Kim et al's modified t-test}
\usage{
kim.t(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
\item{t3}{The value of the modified t statistic.}
\item{p.value}{The p-value for the test.}
}
\description{
Performs a two sample t-test based on Kim et al's modified t statistic.
}
\examples{
x <- c(120,212,NA,300,323,NA,NA,96,178,164)
y <- c(NA,153,304,177,NA,90,89,200,110,40)
kim.t(x,y,alternative = "two.sided") # t3 = 1.141308 p.value = 0.2537419
kim.t(x,y,alternative = "greater") # t3 = 1.141308 p.value = 0.1268709
kim.t(x,y,alternative = "less") # t3 = 1.141308 p.value = 0.8731291
}
