% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liptak.z.R
\name{liptak.z}
\alias{liptak.z}
\title{Liptak's weighted Z-test}
\usage{
liptak.z(x, y, alternative.pool = "greater", alternative.t = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative.pool}{a character string specifying the alternative hypothesis for the weighted Z-test, must be one of "greater" (default), "less", or "two.sided".}

\item{alternative.t}{a character string specifying the alternative hypothesis in the event of a standard two sample t-test, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
\item{z1i}{the value of the weighted z1i statistic.}
\item{z2i}{the value of the weighted z2i statistic.}
\item{p.value}{the p-value for the test.}
}
\description{
Performs a two sample weighted Z-test based on a pooled p-value.
}
\examples{
x <- c(120,212,NA,300,323,NA,NA,96,178,164)
y <- c(NA,153,304,177,NA,90,89,200,110,40)
liptak.z(x,y,alternative.pool = "greater") # z1i = 1.113677 z2i = 0.4201559 p.value = 0.124683
liptak.z(x,y,alternative.pool = "less") # z1i = -1.113677 z2i = -0.4201559 p.value = 0.875317
liptak.z(x,y,alternative.pool = "two.sided") # z1i = 0.6267314 z2i = -0.4520169 pc = 0.4010029 p.value = 0.8020059
}
