% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.het.R
\name{MLE.het}
\alias{MLE.het}
\title{Lin and Stivers's MLE based test under heteroscedasticity}
\usage{
MLE.het(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
\item{z.LS}{The value of the MLE based test statistic under heteroscedasticity.}
\item{df}{degrees of freedom for the test.}
\item{p.value}{The p-value for the test.}
}
\description{
Performs a two sample test based on a Student's t Distribution and an MLE test statistic under heteroscedasticity.
MLE.het(x,y,alternative = c('two.sided','greater','less'))
}
\examples{
# Generating heteroscedastic data:
set.seed(123)
n <- 15
x1 <- 1:n
sd1 <- runif(n,0,5)
het1 <- rnorm(n,0,sd1*x1)
y1 <- x1+het1
y1[c(1,2,3,6)] <- NA

set.seed(321)
x2 <- 1:n
sd2 <- runif(n,0,5)
het2 <- rnorm(n,0,sd2*x2)
y2 <- x2+het2
y2[c(10,13,14,15)] <- NA

MLE.het(y1,y2,alternative = "two.sided") # Z.LS = 0.7982896 p.value = 0.4509315
MLE.het(y1,y2,alternative = "greater") # Z.LS = 0.7982896 p.value = 0.2254657
MLE.het(y1,y2,alternative = "less") # Z.LS = 0.7982896 p.value = 0.7745343
}
