% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.hom.R
\name{MLE.hom}
\alias{MLE.hom}
\title{Ekbohm's MLE based test under homoscedasticity}
\usage{
MLE.hom(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
\item{Z.E}{The value of the MLE based test statistic under homoscedasticity.}
\item{df}{degrees of freedom for the test.}
\item{p.value}{The p-value for the test.}
}
\description{
Performs a two sample test based on a Student's t Distribution and an MLE test statistic under homoscedasticity.
MLE.hom(x,y,alternative = c('two.sided','greater','less'))
}
\examples{
# Generating homoscedastic data:
set.seed(123)
n <- 15
x1 <- 1:n
het1 <- rnorm(n,0,5)
y1 <- x1+het1
y1[c(1,2,3,6)] <- NA

set.seed(321)
x2 <- 1:n
het2 <- rnorm(n,0,5)
y2 <- x2+het2
y2[c(10,13,14,15)] <- NA

MLE.hom(y1,y2,alternative = "two.sided") # Z.E = -0.2249877 p.value = 0.8284157
MLE.hom(y1,y2,alternative = "greater") # Z.E = -0.2249877 p.value = 0.5857922
MLE.hom(y1,y2,alternative = "less") # Z.E = -0.2249877 p.value = 0.4142078
}
