% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.corrected.R
\name{z.corrected}
\alias{z.corrected}
\title{Looney and Jones's corrected z-test}
\usage{
z.corrected(x, y, alternative = "two.sided")
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{a (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less".}
}
\value{
\item{z.stat}{The value of the corrected z statistic.}
\item{p.value}{The p-value for the test.}
}
\description{
Performs a two sample z-test based on a modified variance estimation of the standard z-test by accounting for the correlation among the n1 matched pairs.
z.corrected(x,y,alternative = c('two.sided','greater','less'))
}
\examples{
x <- c(120,212,NA,300,323,NA,NA,96,178,164)
y <- c(NA,153,304,177,NA,90,89,200,110,40)
z.corrected(x,y,alternative = "two.sided") # z.stat = 1.252635 p.value = 0.2103387
z.corrected(x,y,alternative = "greater") # z.stat = 1.252635 p.value = 0.1051694
z.corrected(x,y,alternative = "less") # z.stat = 1.252635 p.value = 0.8948306
}
